fix(consensus): widen cumulative supply accumulators to u128 (#342)
* fix(consensus): widen cumulative supply accumulators to u128

Phase-1 emission (~1.22e21 picocredits ≈ 1.22B BTH) exceeds u64::MAX
(~1.84e19 pico ≈ 18.4M BTH). With overflow-checks=false in release, the
u64 supply accumulators wrapped silently once cumulative emission passed
~18.4M BTH. Because total_mined feeds calculate_block_reward(), a wrapped
value corrupted the emission schedule itself and diverged nodes.

Widen the cumulative monetary accumulators to u128:
- ChainState.total_mined / total_fees_burned
- EmissionController.total_emitted / total_burned (in-memory mirror
  restored from chain state via from_chain_state)
- MintingWork.total_minted (passed to calculate_block_reward)
- calculate_block_reward's total_supply param, plus a u128 reimplementation
  of the tail-reward formula (kept bit-for-bit identical to the cluster-tax
  u64 version) so Phase-2 rewards are correct at real picocredit supply.

Persistence: META_TOTAL_MINED / META_FEES_BURNED now serialize as 16-byte
LE (was 8). This is a consensus-state format change that rides the planned
testnet reset (#323); no migration path is needed.

RPC: handle_chain_info / handle_supply_info now emit totalMined,
totalFeesBurned and circulatingSupply as decimal strings (u128 always
exceeds JS's 2^53 safe-integer limit).

Amounts stay u64: TxOutput.amount, MintingTx.reward and fee are unchanged,
so block/tx/gossip wire format is byte-for-byte identical (asserted).

Adds regression tests: accumulation past u64::MAX (debug+release),
correct calculate_block_reward beyond 2^64 pico, 16-byte metadata
persist/reload, ChainState snapshot serde round-trip > u64::MAX, and a
wire-format guard that monetary fields remain u64-sized.

Closes #333

* fix(fuzz): update fuzz targets for u128 supply accumulators (#333)

The #333 u128 widening of the chain-state supply accumulators
(total_mined, total_fees_burned) missed the workspace-excluded fuzz/
crate, breaking the Fuzz Build Guard (PR):

- fuzz_add_block.rs: `prev.total_mined + expected_reward` no longer
  compiles (u128 + u64); cast the u64 reward to u128. Drop the now
  redundant `as u128` casts on total_mined / total_fees_burned (both
  already u128) so clippy stays green. lottery_pool stays u64, so its
  cast is preserved.
- fuzz_cluster_tax_math.rs: calculate_block_reward's total_supply param
  is now u128; cast the fuzzer's u64 supply at the call site.

No accumulator is narrowed with `as u64`. Verified `cargo fuzz build`
compiles every target clean on the pinned nightly-2025-12-03 toolchain,
and the six #333 regression tests still pass.


---------

+++ fuzz/fuzz_targets/fuzz_cluster_tax_math.rs
@@ -161,7 +161,7 @@ fuzz_target!(|m: FuzzMath| {
 
     // --- 4. Emission / reward schedule -----------------------------------
     // calculate_block_reward must never panic for any height/supply.
-    let _reward = calculate_block_reward(m.height, m.total_supply);
+    let _reward = calculate_block_reward(m.height, m.total_supply as u128);
 
     let policy = MonetaryPolicy::default();
     // Tail reward: pure u128-staged math, must not panic.
